Réécriture du fonctionnement du bundle

+++ Dywee/CMSBundle/Controller/PageController.php
@@ -7,47 +7,30 @@ use Dywee\CMSBundle\Entity\PageElement;
 use Dywee\CMSBundle\Entity\PageStat;
 use Dywee\CMSBundle\Entity\PageTextElement;
 use Dywee\CMSBundle\Form\PageType;
+use Dywee\ModuleBundle\Entity\FormResponseContainer;
+use Dywee\NotificationBundle\Entity\Notification;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\Config\Definition\Exception\Exception;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
 
 class PageController extends Controller
 {
-    public function indexAction()
+    public function indexAction(Request $request)
     {
         $pr = $this->getDoctrine()->getManager()->getRepository('DyweeCMSBundle:Page');
         $page = $pr->findHomePage($this->container->getParameter('website.id'));
 
-        if($page == -1)
+        if($page == null)
             return $this->redirect($this->generateUrl('dywee_install'));
 
-        return $this->viewPageAction($page);
-    }
-
-    public function viewPageAction($page)
-    {
-        $type = $page['type'];
-
-        switch($type)
-        {
-            case 1: return $this->renderHomeAction($page);
-            case 3: return $this->redirect($this->generateUrl('dywee_message_new'));
-            case 4: return $this->render('DyweeNewsBundle:CMS:page.html.twig', array('page' => $page));
-            case 5: return $this->render('DyweeCMSBundle:CMS:view.html.twig', array('page' => $page));
-            case 6: return $this->forward('DyweeEshopBundle:Eshop:pageHandler', array('page' => $page));
-            case 7: return $this->render('DyweeBlogBundle:Blog:page.html.twig', array('page' => $page));
-            case 8: return $this->forward('DyweeModuleBundle:Form:page', array('page' => $page));
-            case 9: return $this->forward('DyweeFaqBundle:Faq:page', array('page' => $page));
-            case 10: return $this->forward('DyweeFaqBundle:PictureGallery:page', array('page' => $page));
-            case 11: return $this->forward('DyweeFaqBundle:VideoGallery:page', array('page' => $page));
-            case 12: return $this->forward('DyweeModuleBundle:MusicGallery:page', array('page' => $page));
-            default: return $this->render('DyweeCMSBundle:CMS:view.html.twig', array('page' => $page));
-        }
+        return $this->viewAction($page->getId(), $request);
     }
 
     public function inMenuSwitchAction(Page $page)
     {
+        if($page->getWebsite()->getId() != $this->get('session')->get('activeWebsite')->getId())
+            throw $this->createAccessDeniedException('Vous ne pouvez pas accéder à cette page');
+
         $em = $this->getDoctrine()->getManager();
 
         $page->setInMenu(!$page->getInMenu());
@@ -65,7 +48,7 @@ class PageController extends Controller
         $pror = $em->getRepository('DyweeProductBundle:Product');
 
         $pageStat = new PageStat();
-        $pageStat->setPage($pr->findOneById($page['id']));
+        $pageStat->setPage($page);
 
         $em->persist($pageStat);
         $em->flush();
@@ -77,6 +60,9 @@ class PageController extends Controller
 
     public function adminViewAction(Page $page)
     {
+        if($page->getWebsite()->getId() != $this->get('session')->get('activeWebsite')->getId())
+            throw $this->createAccessDeniedException('Vous ne pouvez pas accéder à cette page');
+
         $em = $this->getDoctrine()->getManager();
 
         $psr = $em->getRepository('DyweeCMSBundle:PageStat');
@@ -87,22 +73,137 @@ class PageController extends Controller
     public function tableAction()
     {
         $pr = $this->getDoctrine()->getManager()->getRepository('DyweeCMSBundle:Page');
-        $ps = $pr->findByLvl(0);
+        $activeWebsite = $this->get('session')->get('activeWebsite');
+        $ps = $pr->findBy(array('website' => $activeWebsite));
+
         return $this->render('DyweeCMSBundle:CMS:table.html.twig', array('pageList' => $ps));
     }
 
-    public function viewAction($data)
+    public function viewAction($data, Request $request)
+    {
+        /*
+         * Le problème a étudier est celui des pages comprenant un ou plusieurs formulaire(s)
+         * Un render depuis une vue twig rend correctement le formulaire mais ne permet pas de le valider
+         * Le formulaire soumis est considéré comme vide
+         *
+         * Sur les forums on ne trouve pas grand chose vu que c'est pas un use case habituel
+         *
+         * Piste:
+         * Possible que dans l'ancien système et tous les forward etc... de controlleur, on en passait pas les Request à tous
+         * et que ça foutait le bordel parce qu'il recevait un mauvais Request?
+         *
+         * Pourquoi ne pas traiter les formulaires en ajax?
+         */
+        
+        $pageRepository = $this->getDoctrine()->getManager()->getRepository('DyweeCMSBundle:Page');
+
+        if(is_numeric($data))
+            $page = $pageRepository->findById($this->container->getParameter('website.id'), $data);
+        else $page = $pageRepository->findBySeoUrl($this->container->getParameter('website.id'), $data);
+
+        $data = array('page' => $page);
+
+        if($page->hasForm())
+        {
+            $formsId = $page->getForms();
+
+            $em = $this->getDoctrine()->getManager();
+            $formRepository = $em->getRepository('DyweeModuleBundle:DyweeForm');
+
+            $formBuilderService = $this->get('dywee_form.builder');
+
+            $forms = array();
+            foreach($formsId as $formId)
+            {
+                $customForm = $formRepository->findOneBy(array('id' => $formId, 'website' => $this->container->getParameter('website.id')));
+                if($customForm)
+                {
+                    //Le form n'est pas un form à proprement parler mais un objet de type DyweeForm
+                    $formBuilder = $formBuilderService->buildFormBuilder($customForm);
+
+                    //Génération du formulaire
+                    $form = $formBuilder->getForm();
+
+                    //Traitement des formulaires
+                    if($form->handleRequest($request)->isValid())
+                    {
+                        $websiteRepository = $em->getRepository('DyweeWebsiteBundle:Website');
+                        $website = $websiteRepository->findOneById($this->container->getParameter('website.id'));
+
+                        $response = new FormResponseContainer();
+                        $response->setFromForm($customForm, $form->getData());
+
+                        $notification = new Notification();
+                        $notification->setContent('Une nouvelle réponse a été validée pour le formulaire');
+                        $notification->setArgument1($customForm->getId());
+                        $notification->setBundle('module');
+                        $notification->setType('form.response.new');
+                        $notification->setRoutingPath('dywee_customFormResponse_table');
+                        $notification->setRoutingArguments('{id: '.$customForm->getId().'}');
+                        $notification->setWebsite($website);
+
+                        $em->persist($response);
+
+                        $em->flush();
+
+                        $notification->setArgument2($response->getId());
+
+                        $em->persist($notification);
+
+                        $em->flush();
+                    }
+
+                    //Rendu des formulaires
+                    $forms['form_' . $formId] = $form->createView();
+                }
+            }
+
+            //On passe les formulaires à la vue
+            $data['forms'] = $forms;
+        }
+        return $this->render('DyweeCMSBundle:CMS:view.html.twig', $data);
+    }
+
+    /*
+     * Deprecated
+     *
+    public function viewPageAction($page)
+    {
+        $type = $page->getType();
+
+        switch($type)
+        {
+            case 1: return $this->renderHomeAction($page);
+            case 3: return $this->redirect($this->generateUrl('dywee_message_new'));
+            case 4: return $this->render('DyweeNewsBundle:CMS:page.html.twig', array('page' => $page));
+            case 5: return $this->forward('DyweeModuleBundle:Event:page', array('page' => $page));
+            case 6: return $this->forward('DyweeEshopBundle:Eshop:pageHandler', array('page' => $page));
+            case 7: return $this->render('DyweeBlogBundle:Blog:page.html.twig', array('page' => $page));
+            case 8: return $this->forward('DyweeModuleBundle:Form:page', array('page' => $page));
+            case 9: return $this->forward('DyweeFaqBundle:Faq:page', array('page' => $page));
+            case 10: return $this->forward('DyweeFaqBundle:PictureGallery:page', array('page' => $page));
+            case 11: return $this->forward('DyweeFaqBundle:VideoGallery:page', array('page' => $page));
+            case 12: return $this->forward('DyweeModuleBundle:MusicGallery:page', array('page' => $page));
+            default: return $this->render('DyweeCMSBundle:CMS:view.html.twig', array('page' => $page));
+        }
+    }
+     */
+
+    /*
+     * Deprecated
+     */
+    /*public function viewAction($data)
     {
         $em = $this->getDoctrine()->getManager();
         $pr = $em->getRepository('DyweeCMSBundle:Page');
         if(is_numeric($data))
-            $page = $pr->findById($data);
-        else $page = $pr->findBySeoUrl($data);
+            $page = $pr->findById($this->container->getParameter('website.id'), $data);
+        else $page = $pr->findBySeoUrl($this->container->getParameter('website.id'), $data);
 
         if($page)
         {
             $pageStat = new PageStat();
-            $pageStat->setPage($pr->findOneById($page['id']));
+            $pageStat->setPage($page);
 
             $em->persist($pageStat);
             $em->flush();
@@ -110,13 +211,19 @@ class PageController extends Controller
             return $this->viewPageAction($page);
         }
         throw $this->createNotFoundException('Page introuvable');
-    }
+    }*/
 
+    /*
+     * Deprecated
+     */
+    /*
     public function viewContentAction($data)
     {
+        $activeWebsite = $this->container->getParameter('website.id');
+
         $em = $this->getDoctrine()->getManager();
         $pr = $em->getRepository('DyweeCMSBundle:Page');
-        $param = is_numeric($data)?array('id' => $data):0;
+        $param = is_numeric($data)?array('id' => $data, 'website' => $activeWebsite):array('seoUrl' => $data, 'activeWebsite' => $activeWebsite);
         $page = $pr->findOneBy($param);
 
         $pageStat = new PageStat();
@@ -126,23 +233,24 @@ class PageController extends Controller
         $em->flush();
 
         return $this->render('DyweeCMSBundle:CMS:viewContent.html.twig', array('page' => $page));
-    }
+    }*/
 
     public function addAction(Request $request)
     {
-        $em = $this->getDoctrine()->getManager();
-
         $page = new Page();
+
+        $em = $this->getDoctrine()->getManager();
+        $websiteRepository = $em->getRepository('DyweeWebsiteBundle:Website');
+
+        $website = $websiteRepository->findOneById($this->get('session')->get('activeWebsite')->getId());
+
+        $page->setWebsite($website);
 
         $form = $this->get('form.factory')->create(new PageType(), $page);
 
         if($form->handleRequest($request)->isValid())
         {
-            $page->setUpdatedBy($this->get('security.token_storage')->getToken()->getUser());
-
-            $wr = $em->getRepository('DyweeWebsiteBundle:Website');
-            $website = $wr->findOneById(1);
-            $page->setWebsite($website);
+            $page->setUpdatedBy($this->getUser());
 
             $em->persist($page);
             $em->flush();
@@ -156,6 +264,9 @@ class PageController extends Controller
 
     public function updateAction(Page $page, Request $request)
     {
+        if($page->getWebsite()->getId() != $this->get('session')->get('activeWebsite')->getId())
+            throw $this->createAccessDeniedException('Vous ne pouvez pas accéder à cette page');
+
         $form = $this->get('form.factory')->create(new PageType(), $page);
 
         if($form->handleRequest($request)->isValid())
@@ -174,6 +285,9 @@ class PageController extends Controller
 
     public function deleteAction(Page $page)
     {
+        if($page->getWebsite()->getId() != $this->get('session')->get('activeWebsite')->getId())
+            throw $this->createAccessDeniedException('Vous ne pouvez pas accéder à cette page');
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($page);
         $em->flush();
